240629 11:13 main이 될 파일들. 모두 merge를 하였고 Tab1, Tab2 기틀을 잡았음.
