+++ app/src/main/res/layout/activity_second.xml
@@ -1,52 +1,60 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
+<layout>
+    <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
 
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="250dp"
-        android:layout_height="wrap_content"
-        android:text="갤러리"
-        android:textSize="50sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.496"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.453" />
+        <TextView
+            android:id="@+id/textView"
+            android:layout_width="250dp"
+            android:layout_height="wrap_content"
+            android:text="갤러리"
+            android:textSize="50sp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.496"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintVertical_bias="0.453" />
 
-    <Button
-        android:id="@+id/change21Button"
-        android:layout_width="137sp"
-        android:layout_height="wrap_content"
-        android:text="연락처"
-        android:background="@drawable/color_change"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/change22Button" />
+        <Button
+            android:id="@+id/change21Button"
+            android:layout_width="137sp"
+            android:layout_height="wrap_content"
+            android:text="연락처"
+            android:background="@drawable/color_change"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toStartOf="@+id/change22Button" />
 
-    <Button
-        android:id="@+id/change22Button"
-        android:layout_width="138sp"
-        android:layout_height="wrap_content"
-        android:text="갤러리"
-        android:textColor="@color/black"
-        android:backgroundTint="@color/white"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/change21Button"
-        app:layout_constraintEnd_toStartOf="@+id/change23Button" />
+        <Button
+            android:id="@+id/change22Button"
+            android:layout_width="138sp"
+            android:layout_height="wrap_content"
+            android:text="갤러리"
+            android:textColor="@color/black"
+            android:backgroundTint="@color/white"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/change21Button"
+            app:layout_constraintEnd_toStartOf="@+id/change23Button" />
 
-    <Button
-        android:id="@+id/change23Button"
-        android:layout_width="137sp"
-        android:layout_height="wrap_content"
-        android:text="자유"
-        android:background="@drawable/color_change"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/change22Button"
-        app:layout_constraintEnd_toEndOf="parent"/>
+        <Button
+            android:id="@+id/change23Button"
+            android:layout_width="137sp"
+            android:layout_height="wrap_content"
+            android:text="자유"
+            android:background="@drawable/color_change"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintStart_toEndOf="@+id/change22Button"
+            app:layout_constraintEnd_toEndOf="parent"/>
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/recyclerview"
+            android:layout_width="match_parent"
+            app:layout_constraintTop_toBottomOf="@id/change22Button"
+            android:layout_height="wrap_content" />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
