+++ app/src/main/res/layout/item_rv_horizontal.xml
@@ -21,6 +21,7 @@
         android:layout_height="wrap_content"
         android:layout_margin="5dp"
         android:layout_weight="2"
+
         android:textColor="@color/black"
         android:textSize="14sp"
         android:textStyle="bold" />
